Introduce Communication Diagram
Prior to this change, sequence diagrams are great but sometimes you just
want a UML communication diagram.

This change introduces the communication diagram type, which is largely
copied from the sequence diagram type which many elements removed.  A
[sequence expression] is introduced to the signal rule and the different
arrow types are removed.

Some steps have been taken to build a graph of actors, which is simliar
to the class diagram type.  Future changes will remove the linear layout
of actors and render the relationships with messages.

[sequence expression]: https://www.omg.org/spec/UML/2.5.1/PDF#%5B%7B%22num%22%3A2501%2C%22gen%22%3A0%7D%2C%7B%22name%22%3A%22XYZ%22%7D%2C121.5%2C400.3%2C0%5D